# internal/docs/swagger.yaml
definitions:
  entity.ChatMessagesResponse:
    properties:
      messages:
        items:
          properties:
            chat_id:
              type: integer
            content:
              type: string
            message_id:
              type: integer
            message_type:
              type: string
            sender:
              type: integer
          type: object
        type: array
    type: object
...
  entity.CreateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  entity.CreateGroupResponse:
    properties:
      description:
        type: string
      group_id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
...
  entity.NotificationsResponse:
    properties:
      notifications:
        items:
          properties:
            chat_id:
              type: integer
            chat_type:
              type: string
            latest_message:
              type: string
            latest_sender:
              type: integer
            total_messages_count:
              type: integer
          type: object
        type: array
    type: object
...
  entity.ResponseWithStatus:
    properties:
      status:
        type: boolean
    type: object
...
  entity.UpdateGroupColumns:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      group_id:
        type: integer
    type: object
...
  entity.UpdateGroupResponse:
    properties:
      description:
        type: string
      group_id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
...
  errors.Error:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/chat-messages:
    get:
      consumes:
      - application/json
      description: This API for getting chat messages with chat_id
      parameters:
      - description: Chat ID
        in: query
        name: id
        required: true
        type: integer
      - description: Chat Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ChatMessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get Chat Messages
      tags:
      - chat
  /v1/delete-chat-notifications:
    delete:
      consumes:
      - application/json
      description: This API for deleting chats notifications for one user
      parameters:
      - description: Chat ID
        in: query
        name: id
        required: true
        type: integer
      - description: Chat Type
        in: query
        name: type
        required: true
        type: string
      - description: Read Message Count
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseWithStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete Chat Notifications
      tags:
      - chat
  /v1/delete-message/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting a message in chat
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseWithStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete Message
      tags:
      - chat
...
  /v1/get-notifications:
    get:
      consumes:
      - application/json
      description: This API for getting all chats notifications for one user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.NotificationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get All Notifications
      tags:
      - chat
  /v1/group:
    patch:
      consumes:
      - application/json
      description: This API for updating a group columns
      parameters:
      - description: Update Group Columns Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateGroupColumns'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update Group Columns
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: This API for creating a new group
      parameters:
      - description: Create Group Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.CreateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create Group
      tags:
      - chat
    put:
      consumes:
      - application/json
      description: This API for updating a group
      parameters:
      - description: Update Group Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Update Group
      tags:
      - chat
  /v1/group/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting a group
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.DeleteGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete Group
      tags:
      - chat
    get:
      consumes:
      - application/json
      description: This API for getting group with id
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get Group
      tags:
      - chat
  /v1/group/add-user:
    post:
      consumes:
      - application/json
      description: This API for adding a new user to group
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Group ID
        in: query
        name: group_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.ResponseWithStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Add User to Group
      tags:
      - chat
  /v1/group/delete-chat:
    delete:
      consumes:
      - application/json
      description: This API for deleting chat with id
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseWithMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete Chat
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"